v1.3.1
Quick Refactor to move all resizing to a separate file and subroutine to reduce the number of copies of the code.
Code in previous versions has been updated
