+++ MCCMD.py
@@ -7,11 +7,11 @@
 
 # imports the libaries used within Totems+ 
 import os, shutil, getpass, PySimpleGUI as sg
-from PIL import Image, ImageSequence
 from DOC import DOC
 from FUNC import FUN
 from ADVA import ADV
 from ANIM import ANI
+from RESIZE import RES
 
 # outlines the versions and there pack formats
 resourcepackFormat4 = ["1.14","1.14.1","1.14.2","1.14.3","1.14.4"]
@@ -121,61 +121,7 @@ def CMD(textureList, version):
                 if values['inc-orig'] == True:
                     textureList.append("img/totem_of_undying.png")
                 
-                # if the directory resized already exists, then delete it
-                if os.path.exists('C:/Users/' + getpass.getuser() + '/AppData/Roaming/Totems+/resized'):
-                    shutil.rmtree('C:/Users/' + getpass.getuser() + '/AppData/Roaming/Totems+/resized')
-
-                # make the resized folder
-                os.mkdir("C:/Users/" + getpass.getuser() + "/AppData/Roaming/Totems+/resized")
-
-                # new list to hold the new file names
-                pathList = []
-
-                # the path for the resized folder
-                targetIMG = "C:/Users/" + getpass.getuser() + "/AppData/Roaming/Totems+/resized"
-
-                # for loop for each image
-                for i in range(len(textureList)):
-
-                    # finds only the names
-                    splitList = textureList[i].split("/")
-                    # makes a new file name
-                    fileName1 = "resize_" + str(splitList[-1]) 
-
-                    # Run indented code if it is a GIF
-                    if textureList[i].endswith(".gif") == True:
-                        origIMG = Image.open(textureList[i])
-                        size = 128, 128
-                        frames = ImageSequence.Iterator(origIMG)
-                        def thumbnails(frames):
-                            for frame in frames:
-                                thumbnail = frame.copy()
-                                thumbnail = thumbnail.resize(size)
-                                thumbnail.thumbnail(size, Image.ANTIALIAS)
-                                yield thumbnail
-                        frames = thumbnails(frames)
-
-                        resizedGIF = next(frames)
-                        resizedGIF.info = origIMG.info
-                        resizedGIF.save(fileName1, save_all=True, append_images=list(frames), loop=0)
-
-
-                    # Run Code if it is not a GIF
-                    elif textureList[i].endswith(".gif") == False:
-
-                        # copies the images to the resized folder
-                        shutil.copy(textureList[i], targetIMG)
-                        IMGpath = targetIMG + "/" + splitList[-1]
-
-                        # opens and resizes the image, saves with a new name
-                        openIMG = Image.open(IMGpath)
-                        resizedIMG = openIMG.resize((128,128))
-                        resizedIMG.save(fileName1)
-
-                    # moves to the resized folder and saves the path to the path list 
-                    shutil.move(fileName1, targetIMG)
-                    fileName2 = targetIMG + "/" + fileName1
-                    pathList.append(fileName2)
+                pathList = RES(textureList)
 
                 # updates window elements to either be enabled or disabled
                 window.Element('worldBrowse').update(disabled=True)
